move code from records into separate model, queryset, and utils files
