+++ commodities/querysets.py
@@ -1,5 +1,5 @@
 from common.models.mixins.validity import ValidityStartQueryset
-from common.models.records import TrackedModelQuerySet
+from common.models.trackedmodel_queryset import TrackedModelQuerySet
 
 
 class GoodsNomenclatureIndentQuerySet(ValidityStartQueryset, TrackedModelQuerySet):
